backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-filter==23.5
django-jalali==6.0.1
djangorestframework==3.14.0
idna==3.6
jdatetime==4.1.1
Markdown==3.5.2
pillow==10.2.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
